move auto-complete components to own folder
