+++ src/tests/utils/query-input-util.spec.ts
@@ -1,4 +1,4 @@
-import { cleanUpSelectedSuggestion } from '../../app/views/query-runner/query-input/util';
+import { cleanUpSelectedSuggestion } from '../../app/views/query-runner/query-input/auto-complete/util';
 
 
 describe('Query input util should', () => {
